remove templates
