+++ app/javascript/cp/components/ddahEditor.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { DdahForm } from './ddahForm2.js';
+import { DdahForm } from './ddahForm.js';
 import { DdahSpreadsheet } from './ddahSpreadsheet.js';
 import { mockDdahData } from '../../mock_ddah_data.js';  // temporary  .
