--- db/migrate/20180827205932_create_tasks_templates.rb
@@ -1,7 +1,0 @@
-class CreateTasksTemplates < ActiveRecord::Migration[5.1]
-  def change
-    create_join_table :tasks, :templates do |t|
-      t.index [:task_id, :template_id]
-    end
-  end
-end
